Use pipenv for loading collection exercises and events
If a developer isn't using python 3.6.0 or a python version with out
datetime module as their global python version the make file will use
that python version. It will cause the load scripts to fail.

Wrapping the command with pipenv will use the correct version of python
given pyenv is installed