Removing clean database script (#125)
We also have the script reset_database.py to reset the database to a
clean state. For consistency it would be best to only have to maintain
one way of clearing the database down for the tests.

Longer term we need to consider writing tests in such a way that they
will set up their own data and not depend on data being set up already.