Attempt to fix error creating Chrome
We occasionally see failures starting chrome with the error
`Exception ConnectionResetError: [Errno 104] Connection reset by peer`.
There seems to be an open bug around this on github https://github.com/SeleniumHQ/selenium/issues/5296
and as a work around adding a retry mechanism to hopefully ignore it
until it's fixed. Also updating libraries in the chance that might fix
it.